[WIP] Major refactor. Rename lib to globus
